Create My first web page
It is my first web page. I was 14 years old and after I learned HTML language I tried to show my experience to my IT teacher and my classmates and created this web site. And this web site was the beggining of my path.( content of this page is in Ukrainian language)